Add debug logging and refine transcription processing in service

+++ src/api/backend/schemas.py
@@ -2,8 +2,8 @@ from pydantic import BaseModel, Field
 
 
 class Transcription(BaseModel):
-    start: float
-    end: float
+    start: int
+    end: int
     text: str
